requirements.txt
snowflake-connector-python==3.6.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-functions==1.18.0
dbt-core==1.7.0
dbt-snowflake==1.7.0
pandas==2.1.4
pyarrow==14.0.1
pyyaml==6.0.1
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
numpy==1.26.0
